+++ app/users-register/content-page.php
@@ -6,10 +6,10 @@ $resp->reqPathComponents = \xan\paramEncode( $aloe_request->path_components_get(
 $resp->reqPost = \xan\paramEncode( $aloe_request->post );
 // $resp->reqID = $resp->reqPathComponents[ 1 ];
 $resp->moduleName = $mmUsersRegister->NameModule;
-$resp->headTitle = $mmUsersRegister->NameModule;
+$resp->headTitle = $mmUsersRegister->NamePlural;
 $resp->headLogoutAuto = false;
 $resp->navInclude = false;
-$resp->contentHeader = $mmUsersRegister->FontAwesome . STR_NBSP . $mmUsersRegister->NameModule;
+$resp->contentHeader = $mmUsersRegister->FontAwesome . STR_NBSP . $mmUsersRegister->NamePlural . STR_NBSP;
 
 ///////////////////////////////////////////////////////////
 // Content Load Now